refact: base zoom assertions on unclippedBoundsMatcher
