+++ feature/gallery/src/androidTest/java/com/jeanbarrossilva/orca/feature/gallery/SemanticsNodeInteractionExtensionsTests.kt
@@ -23,59 +23,15 @@ import androidx.compose.ui.test.assertContentDescriptionEquals
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onRoot
 import androidx.test.platform.app.InstrumentationRegistry
-import com.jeanbarrossilva.orca.feature.gallery.test.assertIsZoomedIn
-import com.jeanbarrossilva.orca.feature.gallery.test.assertIsZoomedOut
 import com.jeanbarrossilva.orca.feature.gallery.test.onPage
 import com.jeanbarrossilva.orca.feature.gallery.test.onPager
 import com.jeanbarrossilva.orca.feature.gallery.test.performScrollToPageAt
-import com.jeanbarrossilva.orca.feature.gallery.test.performZoomIn
-import com.jeanbarrossilva.orca.feature.gallery.test.performZoomOut
 import com.jeanbarrossilva.orca.platform.ui.component.timeline.post.formatted
 import org.junit.Rule
 import org.junit.Test
 
 internal class SemanticsNodeInteractionExtensionsTests {
   @get:Rule val composeRule = createComposeRule()
-
-  @Test
-  fun performsZoomIn() {
-    composeRule.apply { setContent { Gallery() } }.onPage().performZoomIn().assertIsZoomedIn()
-  }
-
-  @Test(expected = AssertionError::class)
-  fun throwsWhenAssertingThatIsZoomedInWhenItIsNot() {
-    composeRule.apply { setContent { Gallery() } }.onPage().performZoomOut().assertIsZoomedIn()
-  }
-
-  @Test
-  fun assertsIsZoomedIn() {
-    composeRule.apply { setContent { Gallery() } }.onPage().performZoomIn().assertIsZoomedIn()
-  }
-
-  @Test
-  fun performsZoomOut() {
-    composeRule
-      .apply { setContent { Gallery() } }
-      .onPage()
-      .performZoomIn()
-      .performZoomOut()
-      .assertIsZoomedOut()
-  }
-
-  @Test(expected = AssertionError::class)
-  fun throwsWhenAssertingThatIsZoomedOutWhenItIsNot() {
-    composeRule.apply { setContent { Gallery() } }.onPage().performZoomIn().assertIsZoomedOut()
-  }
-
-  @Test
-  fun assertsIsZoomedOut() {
-    composeRule
-      .apply { setContent { Gallery() } }
-      .onPage()
-      .performZoomIn()
-      .performZoomOut()
-      .assertIsZoomedOut()
-  }
 
   @Test(expected = AssertionError::class)
   fun throwsWhenScrollingToPageOfNonGalleryPagerNode() {
